feat combine+mvvm적용
