--- gongjipsa/gongjipsaApp.swift
@@ -1,17 +1,0 @@
-//
-//  gongjipsaApp.swift
-//  gongjipsa
-//
-//  Created by Jihaha kim on 2024/06/25.
-//
-
-import SwiftUI
-
-@main
-struct gongjipsaApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
